Ai pv (#45)
* Crud V-1

* ADD : add feature authentication/verifyEmail

* Merge branch 'dev' of https://github.com/FUBLOG/FUBLOG_BE into dev

* Config Post

* ADD: authentication with accesstoken

* post-ver3

* 22

* Add: message feature and socket connect

* done Tag and report

* Done tag and report

* done Vers4

* Vanmanh (#6)

* done Tag and report

* Done tag and report

* done Vers4

---------


* save

* code

* ?

* ADD : cors and notification feature

* fix ver1

* 1stver_Userinfo

* update

* done

* Feat/authentication (#10)

* 22

* save

* ADD : cors and notification feature

* update

* done

* FIX : Cors

* done

* save commit

* Set Up Socket

* FIX: fix cors socket

* fix ssl cors

* fix ssl cors 2

* 1

* update

* done

* fix

* fix

* fix

* 123

* fix

* 1 so chuc nang ve size cua anh

* ADD: cron-job

* before update

* ...

* FIX: login api

* FIX: login api

* save

* ...

---------

+++ src/services/addFriend.service.js
@@ -4,25 +4,35 @@ const User = require('../model/user.model');
 const { isMongoId } = require('validator');
 const userInfor = require('../model/userInfo.model');
 const addFriendRepo = require('../repository/addFriend.repo')
-
+const {NotFoundError} = require("../core/response/error.response")
+const {CREATED} = require("../core/response/success.response")
+const users = require ("../model/user.model");
+const { log } = require('../logger/log.system');
 
 class addFriendService {
     static sendFriendRequest = async (sourceID, targetID) => {
         if (!isMongoId(sourceID) || !isMongoId(targetID)) {
-            throw new Error('invalid userID');
+            throw new NotFoundError('invalid userID');
         }
     
         const sender = await User.findById(sourceID);
         const receiver = await User.findById(targetID);
     
         if (!sender || !receiver) {
-            throw new Error('user not found');
+            throw new NotFoundError('user not found');
         }
     
         const existingRequest = await addFriendRepo.findRequest(sourceID, targetID);
     
-        if (existingRequest) {
-            throw new Error('already sent');
+        if (existingRequest && existingRequest.status === "pending") {
+            throw new CREATED('already sent');
+        }
+        if (existingRequest && existingRequest.status === 'accepted') {
+            throw new CREATED('You are already friends');
+        }
+
+        if (existingRequest && existingRequest.status === 'declined') {
+            throw new CREATED('Friend request was declined');
         }
     
         const createdRequest = await addFriendRepo.createRequest(sourceID, targetID);
@@ -33,37 +43,55 @@ class addFriendService {
     
     static acceptFriendRequest = async (sourceID, targetID) => {
         if (!isMongoId(sourceID) || !isMongoId(targetID)) {
-            throw new Error('invalid userID');
+            throw new NotFoundError('invalid userID');
         }  
-        const friendRequest = await addFriendRepo.findRequest(sourceID, targetID);   
-        if (!friendRequest ) {
-            throw new Error('friend request not found or already processed');
-        }
-
+        // const existingRequest = await addFriendRepo.findRequest(sourceID, targetID);   
+        // if (!existingRequest || existingRequest.status !== 'pending') {
+        //     throw new NotFoundError('friend request not found or already processed');
+        // }
+    
         const senderInfo = await userInfor.findOne({ user_id: sourceID });
         const receiverInfo = await userInfor.findOne({ user_id: targetID });
-        console.log(senderInfo)
-        console.log(receiverInfo)
-        if (!senderInfo.friendList.includes(targetID)) {
-            senderInfo.friendList.push(targetID);
+        const receiver = await users.findOne({ _id: targetID })
+        const sender = await users.findOne({ _id: sourceID })
+    
+        if (!senderInfo.friendList.find(friend => friend.friend_id.equals(targetID))) {
+            senderInfo.friendList.push({
+                friend_id: targetID,
+                displayName: receiver.displayName,
+                avatar: receiverInfo.avatar,
+                profileHash: receiver.profileHash
+            });
             await senderInfo.save();
         }
     
-        if (!receiverInfo.friendList.includes(sourceID)) {
-            receiverInfo.friendList.push(sourceID);
+        if (!receiverInfo.friendList.find(friend => friend.friend_id.equals(sourceID))) {
+            receiverInfo.friendList.push({
+                friend_id: sourceID,
+                displayName: sender.displayName,
+                avatar: senderInfo.avatar,
+                profileHash: sender.profileHash
+            });
             await receiverInfo.save();
         }
-        const updatedFriendRequest = await addFriendRepo.findRequest(sourceID, targetID);
-
-        return updatedFriendRequest;
-        // return 'friend request accepted';
+        
+        await addFriendRepo.updateRequestStatus(sourceID, targetID, "accepted");
+        await addFriendRepo.findRequest(sourceID, targetID);
+    
+        return {
+            friend_id: targetID,
+            displayName: receiver.displayName,
+            avatar: receiverInfo.avatar,
+            profileHash: receiver.profileHash
+        };
     };
+    
     
     
    
     static getAllFriendRequests = async (userID) => {
         if (!isMongoId(userID)) {
-            throw new Error('invalid userID');
+            throw new NotFoundError('invalid userID');
         }
 
         const requests = await addFriendRepo.findAllRequests(userID);
@@ -71,40 +99,30 @@ class addFriendService {
         return requests;
     };
 
-    static declineFriendRequest = async (sourceID, targetID) => {
+  static declineFriendRequest = async (sourceID, targetID) => {
         if (!isMongoId(sourceID) || !isMongoId(targetID)) {
-            throw new Error('invalid userID');
+            throw new NotFoundError('invalid userID');
         }
-    
+        const existingRequest = await addFriendRepo.findRequest(sourceID, targetID);   
+        if (!existingRequest || existingRequest.status !== 'pending') {
+            throw new NotFoundError('friend request not found or already processed');
+        }
         const deletedRequest = await addFriendRepo.deleteRequest(sourceID, targetID);
-    
+
         return deletedRequest;
-    };static declineFriendRequest = async (sourceID, targetID) => {
-    if (!isMongoId(sourceID) || !isMongoId(targetID)) {
-        throw new Error('invalid userID');
-    }
+        await addFriendRepo.updateRequestStatus(sourceID, targetID, "declined");
 
-    const deletedRequest = await addFriendRepo.deleteRequest(sourceID, targetID);
-
-    return deletedRequest;
-};
-
-    
-    
-static declineFriendRequest = async (sourceID, targetID) => {
-    if (!isMongoId(sourceID) || !isMongoId(targetID)) {
-        throw new Error('invalid userID');
-    }
-
-    const deletedRequest = await addFriendRepo.deleteRequest(sourceID, targetID);
-
-    return deletedRequest;
-};
-
+        
+    };
 
     static unFriend = async (sourceID, targetID) => {
         if (!isMongoId(sourceID) || !isMongoId(targetID)) {
-            throw new Error('invalid userID');
+            throw new NotFoundError('invalid userID');
+        }
+        const existingRequest = await addFriendRepo.findRequest(sourceID, targetID);
+
+        if (!existingRequest || existingRequest.status === 'pending') {
+            throw new NotFoundError('friend request not found or not a friend');
         }
         const senderInfo = await userInfor.findOne({ user_id: sourceID });
         const receiverInfo = await userInfor.findOne({ user_id: targetID });
@@ -119,30 +137,72 @@ static declineFriendRequest = async (sourceID, targetID) => {
             await receiverInfo.save();
         }
 
+        await addFriendRepo.updateRequestStatus(sourceID, targetID,"declined")
 
         const deletedRequest = await addFriendRepo.deleteRequest(sourceID, targetID);
       
         return deletedRequest;
     };
 
+    // static blockFriend = async (sourceID, targetID) => {
+    //     if (!isMongoId(sourceID) || !isMongoId(targetID)) {
+    //         throw new NotFoundError('invalid userID');
+    //     }
+    //     const existingRequest = await addFriendRepo.findRequest(sourceID, targetID);
+
+    //     if (!existingRequest || existingRequest.status === 'pending') {
+    //         throw new NotFoundError('friend request not found or not a friend');
+    //     }
+    //     const senderInfo = await userInfor.findOne({ user_id: sourceID });
+
+    //     if (!senderInfo.blockList.includes(targetID)) {
+    //         senderInfo.blockList.push(targetID);
+    //         await senderInfo.save();
+    //     }
+    //     await addFriendRepo.updateRequestStatus(sourceID, targetID,"declined")
+
+    //     const deletedRequest = await addFriendRepo.deleteRequest(sourceID, targetID);
+
+    //     return deletedRequest;
+    // };
     static blockFriend = async (sourceID, targetID) => {
-        if (!isMongoId(sourceID) || !isMongoId(targetID)) {
-            throw new Error('invalid userID');
-        }
+        console.log(sourceID);
+    if (!isMongoId(sourceID) || !isMongoId(targetID)) {
+        throw new NotFoundError('invalid userID');
+    }
 
-        const senderInfo = await userInfor.findOne({ user_id: sourceID });
+    const existingRequest = await addFriendRepo.findRequest(sourceID, targetID);
+    // if (!existingRequest || existingRequest.status === 'pending') {
+    //     throw new NotFoundError('friend request not found or not a friend');
+    // }
 
-        if (!senderInfo.blockList.includes(targetID)) {
-            senderInfo.blockList.push(targetID);
-            await senderInfo.save();
-        }
+    const senderInfo = await userInfor.findOne({ user_id: sourceID });
+    const receiverInfo = await userInfor.findOne({ user_id: targetID });
 
-        const deletedRequest = await addFriendRepo.deleteRequest(sourceID, targetID);
+    const sender = await users.findOne({ _id: sourceID })
+    console.log(sender);
+    if (!receiverInfo.blockList.find(friend => friend.friend_id.equals(sourceID))) {
+        receiverInfo.blockList.push({
+            friend_id: sourceID,
+            displayName: sender.displayName,
+            avatar: senderInfo.avatar,
+            profileHash: sender.profileHash
+        });
+        await receiverInfo.save();
+    }
 
-        return deletedRequest;
+    await addFriendRepo.updateRequestStatus(sourceID, targetID, "declined");
+    await addFriendRepo.deleteRequest(sourceID, targetID);
+
+    return {
+        friend_id: sourceID,
+        displayName: sender.displayName,
+        avatar: senderInfo.avatar,
+        profileHash: sender.profileHash
     };
+};
+
 }
 
 
 module.exports = addFriendService;
-
